disable shutdown cleanup
As it turns out, we cannot simply delete empty queues
because there might still be delayed messages for the same queues around.
In that case the delayed messages would be silently dropped.

One way around that would be to inspect the delaying queue
to see which target queues can be safely deleted,
but the overhead seems excessive.

revert 01bc299, 63ef4ae