test: run on default port
Changing the rabbitmq port with RABBITMQ_NODE_PORT works fine for the rabbitmq-server itself,
but rabbitmqctl does not seem to notice the variable.
This leads to incorrect error messages on start-up and too little delay.
So now we're running the server on the default port.
To avoid some of the inherent risks, our tests will make sure they're running in the correct environment
with another env variable in the test container.